implement the cart context and the cart context CartProvider
The CartContext holds the data for the cart as well
the functions to add and remove item in the cart
The CartProvider is a React component in which
I use to wrap all the components I'll need to use the
context. So the whole app.js

In the component, I use the useContext hook to store the
store in a constant and use it in the component as an object.