create the useReducer logic and one action for add items in the cart
the useReducer accepts two arguments, the first is a
reducer function to be called when the hook is called
and the second is the initial state

the function passed to the hook accepts two arguments.
the first is a snapshot of the state and the second is
the action, which is recommended to be an object.