Wired the context and items are being added to the cart.
First, I used the useRefs() hook to create a Ref state
to hold the input number.
Since I used the useRefs() in a custom component, I wrapped
the custom element in a React.forwardRef() to send the ref
from the native input element through the custom Input
element to the upper element by sending ref={ref} in the
native input element.

I stored the input number in a constant in the MealItemForm
component using the normal syntax amountRef.current.value.

I passed the number to the function received through props
from the MealItem component

In the MealItem element, I send the function to the MealItemForm
component through props.
Inside the function, I received the input number as argument
Then, I create an item object and send it to the CartContext
method calling addItem.

The method prepares the action object and performs the adding